+++ tests/env_config/conftest.py
@@ -1,6 +1,7 @@
-from typing import Type, Optional
+from typing import Optional, Type
 
 from pytest import fixture
+
 from datek_app_utils.env_config.base import BaseConfig